Refactored and improved WhatsappWebAutomation.py
The 'CHROME_PROFILE_PATH' variable was defined twice, once in the 'qrcode_auth()' and once in the 'send_message()' functions. Now, it's defined just once in the main() funtcion. Also, the variable now uses the 'os' module to get the user path, so the variable doesn't need manual edting when working in different computers.

The message that will be sent is stored in 'message.txt', so the user doesn't need to edit the code.