Add a config.json file and improve user interaction through terminal
The config.json takes parameters like, what is the name of the file with the phone numbers and within that file what is the phone column. This allows us to make an .exe out of our file since the user doesn't have to edit the code itself.
Print Logo to the terminal and some prompt messages like " PRESS ENTER TO START" and "You may close this window"