Tkinter GUI created and documentation in the code improved.
The main file WhatsappWebAutomation.py is renamed to main.py.
Error handling improved by prompting a message box with the error content in it then finishing execution of app.
Functions are better documented as I tried to add more comments.
An interactive graphic interface checks if teh user is ready to start, whether they want to scan a new QR and logs what the scrip is doing followed by a progressbar and the number of successes and failures.